Move implementations to their respective modules
