--- api/common/src/test/kotlin/com/github/jcornaz/collekt/DefaultPersistentListTest.kt
@@ -1,8 +1,0 @@
-package com.github.jcornaz.collekt
-
-class DefaultPersistentListTest : PersistentListTest() {
-    override val factory = object : PersistentListFactory {
-        override fun <E> empty(): PersistentList<E> = emptyPersistentList<Nothing>()
-        override fun <E> from(elements: Iterable<E>) = elements.toPersistentList()
-    }
-}
