+++ api/common/src/test/kotlin/com/github/jcornaz/collekt/api/DefaultPersistentListTest.kt
@@ -1,0 +1,8 @@
+package com.github.jcornaz.collekt.api
+
+class DefaultPersistentListTest : PersistentListTest() {
+    override val factory = object : PersistentListFactory {
+        override fun <E> empty(): PersistentList<E> = emptyPersistentList<Nothing>()
+        override fun <E> from(elements: Iterable<E>) = elements.toPersistentList()
+    }
+}
