--- api/js/src/main/kotlin/com/github/jcornaz/collekt/DefaultImplementations.kt
@@ -1,5 +1,0 @@
-package com.github.jcornaz.collekt
-
-import com.github.jcornaz.collekt.impl.KotlinList
-
-internal actual val DefaultListFactory: PersistentListFactory = KotlinList
