+++ api/js/src/main/kotlin/com/github/jcornaz/collekt/api/DefaultImplementations.kt
@@ -1,0 +1,3 @@
+package com.github.jcornaz.collekt.api
+
+internal actual val DefaultListFactory: PersistentListFactory = TODO()
