--- benchmark/src/main/kotlin/com/github/jcornaz/collekt/RemoveElementBenchmark.kt
@@ -1,12 +1,0 @@
-package com.github.jcornaz.collekt
-
-import org.openjdk.jmh.annotations.Benchmark
-
-open class RemoveElementBenchmark {
-
-    @Benchmark
-    fun removeIndexInArrayList(input: ArrayListInput) = input.list.removeAt(input.randomIndex)
-
-    @Benchmark
-    fun removeIndex(input: PersistentInput) = input.list.minusIndex(input.randomIndex)
-}
