+++ benchmark/src/main/kotlin/com/github/jcornaz/collekt/api/RemoveElementBenchmark.kt
@@ -1,0 +1,12 @@
+package com.github.jcornaz.collekt.api
+
+import org.openjdk.jmh.annotations.Benchmark
+
+open class RemoveElementBenchmark {
+
+    @Benchmark
+    fun removeIndexInArrayList(input: ArrayListInput) = input.list.removeAt(input.randomIndex)
+
+    @Benchmark
+    fun removeIndex(input: PersistentInput) = input.list.minusIndex(input.randomIndex)
+}
